+++ classes/helper.php
@@ -96,8 +96,7 @@ class local_course_template_helper {
                 get_config('local_course_template', 'templatenameformat'));
 
             // Check if the idnumber is cached.
-            $cache = cache::make('local_course_template', 'templates');
-            $courseid = $cache->get($shortname);
+            $courseid = self::get_cached_course_id($shortname);
             if ($courseid == false) {
                 $course = $DB->get_record('course', array('shortname' => $shortname));
                 if (empty($course)) {
@@ -105,7 +104,7 @@ class local_course_template_helper {
                     $defaultshortname = get_config('local_course_template', 'defaulttemplate');
                     $defaultcourse = $DB->get_record('course', array('shortname' => $defaultshortname));
                     if (!empty($defaultshortname && !empty($defaultcourse))) {
-                        $cache->set($defaultshortname, $defaultcourse->id);
+                        self::set_cached_course_id($defaultshortname, $defaultcourse->id);
                         return $defaultcourse->id;
                     }
                     return false;
@@ -119,6 +118,43 @@ class local_course_template_helper {
             // This course doesn't conform to the given naming convention, so skip.
             return false;
         }
+    }
+
+    /**
+     * Returns the cached template course id, if it exists.
+     *
+     * Returns the cached template course id if it exists, or false it it does not. Also returns
+     * false if caching is disabled.
+     *
+     * @param string $shortname the shortname of the template course
+     * @return int|boolean
+     */
+    private static function get_cached_course_id($shortname) {
+        $enablecaching = get_config('local_course_template', 'enablecaching');
+        if (empty($enablecaching) || $enablecaching == 0) {
+            return false;
+        }
+        $cache = cache::make('local_course_template', 'templates');
+        $courseid = $cache->get($shortname);
+        return $courseid;
+    }
+
+    /**
+     * Set the cached template course id.
+     *
+     * Caches the course id for the given template course. It does nothing
+     * if caching is disabled.
+     *
+     * @param string $shortname the course shortname
+     * @param int $courseid the course id
+     */
+    private static function set_cached_course_id($shortname, $courseid) {
+        $enablecaching = get_config('local_course_template', 'enablecaching');
+        if (empty($enablecaching) || $enablecaching == 0) {
+            return;
+        }
+        $cache = cache::make('local_course_template', 'templates');
+        $cache->set($shortname, $courseid);
     }
 
     /**